refactor: 게임컴포넌트 3D와 HTML 컴포넌트 분리
