+++ src/components/Game/GameContainer/GameContainer.module.css
@@ -1,0 +1,26 @@
+.container {
+  position: relative;
+  width: 100%;
+  height: 100vh;
+}
+
+.canvas {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+}
+
+.uiOverlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  pointer-events: none;
+}
+
+.uiOverlay > * {
+  pointer-events: auto;
+}
