+++ src/components/Game/GameUserInterface/GameUserInterface.module.css
@@ -1,0 +1,14 @@
+.gameUI {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  pointer-events: none;
+  width: 100%;
+  height: 100%;
+}
+
+.gameUI > * {
+  pointer-events: auto;
+}
